requirements.txt
Flask>=2.0.3
Flask_SQLAlchemy>=2.5.1
Flask_Migrate>=3.1.0
python-dotenv>=0.21.0
openai>=1.0.0
PyPDF2>=1.26.0